test: merge frames.spec.ts into heap.spec.ts
These are testing non-leaking behavior.
